Add files via upload
FakeNews is the main file that preprocesses the data and trains different classification models and compares their results. A step by step walkthrough is in the documentation pdf. The TestScript file is the one that assumes that new data is given and predicts the outputs based on the pre-trained model in FakeNews.